SimulatorNFCManager
